+++ org-vscode/out/extension.js
@@ -17,6 +17,7 @@ const changeDirectory = require("./changeDirectory");
 const keywordRight = require("./keywordRight");
 const keywordLeft = require("./keywordLeft");
 const setTodoState = require("./setTodoState");
+const { cyclePriorityForward, cyclePriorityBackward } = require("./cyclePriority");
 const moveUp = require("./moveUp");
 const moveDown = require("./moveDown");
 const getTags = require("./tags");
@@ -310,6 +311,8 @@ function activate(ctx) {
   ctx.subscriptions.push(vscode.commands.registerCommand("extension.toggleStatusRight", keywordRight));
   ctx.subscriptions.push(vscode.commands.registerCommand("extension.toggleStatusLeft", keywordLeft));
   ctx.subscriptions.push(vscode.commands.registerCommand("extension.setTodoState", setTodoState));
+  ctx.subscriptions.push(vscode.commands.registerCommand("org-vscode.cyclePriority", cyclePriorityForward));
+  ctx.subscriptions.push(vscode.commands.registerCommand("org-vscode.cyclePriorityBackward", cyclePriorityBackward));
   ctx.subscriptions.push(vscode.commands.registerCommand("extension.scheduling", scheduling));
   ctx.subscriptions.push(vscode.commands.registerCommand("extension.deadline", deadline));
   ctx.subscriptions.push(vscode.commands.registerCommand("extension.moveBlockUp", moveUp));